scripts/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

argcomplete==3.2.2
beautifulsoup4==4.12.2
chromadb==0.4.19
colorama==0.4.6
faiss-cpu==1.7.4
huggingface-hub==0.19.4
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-experimental==0.0.49
langchain-openai==0.0.3
llama_cpp_python==0.2.26
matplotlib==3.8.2
numpy==1.26.2
openai==1.9.0
pandas==2.1.3
Pillow==10.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pytube==15.0.0
qdrant-client==1.6.9
ragas==0.1.0
requests==2.31.0
streamlit==1.28.2
streamlit-chat==0.1.1
streamlit-extras==0.4.0
torch==2.1.1
transformers==4.35.2
wordcloud==1.9.3
